Bunch of updates
Bunch of stuff updated here! We've now got a date property in Notion if you say "passed on " at the beginning of your discord message (because that's the particular syntax that the place I made this for uses, feel free to change it), and there is duplicate protection inbuilt! (Because I was having some issues with it)